+++ CHANGELOG.md
@@ -30,7 +30,7 @@ You can check your current version with the following command:
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
 ## 2.2.0a20
-**Jan 8, 2025**
+**Jan 9, 2025**
 - Added `update` extension method to `pystac::Item`
 
 ## 2.2.0a19