+++ CHANGELOG.md
@@ -29,6 +29,10 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
+## 2.2.0a20
+**Jan 9, 2025**
+- Added `update` extension method to `pystac::Item`
+
 ## 2.2.0a19
 **Dec 20, 2024**
 - Switched to using stac client descriptor in `Asset` class and reduced duplication.